+++ Resources.cpp
@@ -2666,17 +2666,17 @@ int& Sexy::GetSoundRefById(int theId)
 
 static Sexy::ResourceId GetIdByVariable(const void *theVariable)
 {
-	typedef std::map<int,int> MyMap;
+	typedef std::map<int64_t,int64_t> MyMap;
 	static MyMap aMap;
 	if(gNeedRecalcVariableToIdMap)
 	{
 		gNeedRecalcVariableToIdMap = false;
 		aMap.clear();
 		for(int i=0; i<RESOURCE_ID_MAX; i++)
-			aMap[*(int*)gResources[i]] = i;
+			aMap[*(int64_t*)gResources[i]] = i;
 	}
 
-	MyMap::iterator anItr = aMap.find((int)theVariable);
+	MyMap::iterator anItr = aMap.find((int64_t)theVariable);
 	if (anItr == aMap.end())
 		return RESOURCE_ID_MAX;
 	else
